requirements.txt
Django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9.0
Pillow>=5.3.0,<5.4.0